Rename two directories. Delete default react logos.
