+++ src/components/RandomImage/styles.tsx
