+++ src/routes/Breeds.tsx
@@ -1,5 +1,5 @@
 import BreedList from '../components/breedList/BreedList';
-import Header from '../components/header/Header';
+import Header from '../components/Header/Header';
 
 export default function Breeds() {
     return(
