+++ src/routes/NoMatch.tsx
@@ -1,4 +1,4 @@
-import Header from '../components/header/Header';
+import Header from '../components/Header/Header';
 import Error from '../components/Error/Error';
 
 export default function NoMatch() {
